+++ board/rt1051/crashdump/consoledump.hpp
@@ -1,0 +1,22 @@
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
+// For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
+
+#pragma once
+
+#include <cstdint>
+
+// Exception Stack Frame
+typedef struct __attribute__((packed))
+{
+    uint32_t r0;
+    uint32_t r1;
+    uint32_t r2;
+    uint32_t r3;
+    uint32_t r12;
+    uint32_t lr;
+    uint32_t pc;
+    uint32_t psr;
+#if defined(__ARM_ARCH_7EM__)
+    uint32_t s[16];
+#endif
+} syslog_exception_stack_frame_t;
