+++ module-bsp/board/rt1051/common/startup_mimxrt1052.cpp
@@ -1,3 +1,6 @@
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
+// For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
+
 //*****************************************************************************
 // MIMXRT1052 startup code for use with MCUXpresso IDE
 //
@@ -774,352 +777,6 @@ WEAK_AV void PendSV_Handler(void)
 WEAK_AV void SysTick_Handler(void)
 {
     while (1) {}
-}
-
-#if defined(__cplusplus)
-extern "C"
-{
-#endif
-
-    // Exception Stack Frame
-    typedef struct __attribute__((packed))
-    {
-        uint32_t r0;
-        uint32_t r1;
-        uint32_t r2;
-        uint32_t r3;
-        uint32_t r12;
-        uint32_t lr;
-        uint32_t pc;
-        uint32_t psr;
-#if defined(__ARM_ARCH_7EM__)
-        uint32_t s[16];
-#endif
-    } syslog_exception_stack_frame_t;
-
-    /// Hard Fault Fault Status Register
-    struct __attribute__((packed)) hfsr_t
-    {
-        union
-        {
-            uint32_t all;
-            struct
-            {
-                uint32_t rsrvd0 : 1;
-                uint32_t vecttbl : 1;
-                uint32_t rsrvd1 : 28;
-                uint32_t forced : 1;
-                uint32_t debugevt : 1;
-            };
-        };
-    };
-
-    /// MemManage Fault Status Register
-    struct __attribute__((packed)) mmfsr_t
-    {
-        union
-        {
-            uint8_t all;
-            struct
-            {
-                uint8_t iaccviol : 1;
-                uint8_t daccviol : 1;
-                uint8_t rsrvd2 : 1;
-                uint8_t munstkerr : 1;
-                uint8_t mstkerr : 1;
-                uint8_t mlsperr : 1;
-                uint8_t rsrvd6 : 1;
-                uint8_t mmarvalid : 1;
-            };
-        };
-    };
-
-    /// BusFault Status Register
-    struct __attribute__((packed)) bfsr_t
-    {
-        union
-        {
-            uint8_t all;
-            struct
-            {
-                uint8_t ibuserr : 1;
-                uint8_t preciserr : 1;
-                uint8_t impreciserr : 1;
-                uint8_t unstkerr : 1;
-                uint8_t stkerr : 1;
-                uint8_t lsperr : 1;
-                uint8_t rsrvd6 : 1;
-                uint8_t bfarvalid : 1;
-            };
-        };
-    };
-
-    /// UsageFault Status Register
-    struct __attribute__((packed)) ufsr_t
-    {
-        union
-        {
-            uint16_t all;
-            struct
-            {
-                uint8_t undefinstr : 1;
-                uint8_t invstate : 1;
-                uint8_t invpc : 1;
-                uint8_t nocp : 1;
-                uint8_t rsrvd4_7 : 3;
-                uint8_t unaligned : 1;
-                uint8_t divbyzero : 1;
-                uint8_t rsrvd10_15 : 6;
-            };
-        };
-    };
-
-    /// Configurable Fault Status Register
-    struct __attribute__((packed)) cfsr_t
-    {
-        union
-        {
-            uint32_t all;
-            struct
-            {
-                mmfsr_t mmfsr;
-                bfsr_t bfsr;
-                ufsr_t ufsr;
-            };
-        };
-    };
-
-    /// Auxiliary Bus Fault Status register
-    struct __attribute__((packed)) abfsr_t
-    {
-        union
-        {
-            uint32_t all;
-            struct
-            {
-                uint32_t itcm : 1;
-                uint32_t dtcm : 1;
-                uint32_t ahbp : 1;
-                uint32_t axim : 1;
-                uint32_t eppb : 1;
-                uint32_t rsrvd5_7 : 3;
-                uint32_t aximtype : 2;
-                uint32_t rsrvd10_31 : 1;
-            };
-        };
-    };
-
-    struct __attribute__((packed)) syslog_t
-    {
-        syslog_exception_stack_frame_t stackFrame;
-        struct
-        {
-            hfsr_t hfsr;         /// Hard Fault Fault Status Register
-            cfsr_t cfsr;         /// Configurable Fault Status Register
-            uint32_t mmfar;      /// MemManage Fault Address Register
-            uint32_t bfar;       /// BusFault Address Register
-            abfsr_t abfsr;       /// Auxiliary Bus Fault Status register
-            uint32_t exc_return; /// EXC_RETURN value for the exception being serviced
-        } registers;
-    };
-
-#define DEBUG_PRINT_HARD_FAULT_INFO    1
-#define DEBUG_DETAILED_HARD_FAULT_INFO 1
-
-    static void printHardFaultInfo(const syslog_t &syslog)
-    {
-#if (DEBUG_PRINT_HARD_FAULT_INFO == 1)
-        LOG_FATAL("!!! HardFault Detected !!!");
-        LOG_FATAL("* Stack Frame:");
-        LOG_FATAL("R0  = 0x%0" PRIX32, syslog.stackFrame.r0);
-        LOG_FATAL("R1  = 0x%0" PRIX32, syslog.stackFrame.r1);
-        LOG_FATAL("R2  = 0x%0" PRIX32, syslog.stackFrame.r2);
-        LOG_FATAL("R3  = 0x%0" PRIX32, syslog.stackFrame.r3);
-        LOG_FATAL("R12 = 0x%0" PRIX32, syslog.stackFrame.r12);
-        LOG_FATAL("LR  = 0x%0" PRIX32, syslog.stackFrame.lr);
-        LOG_FATAL("PC  = 0x%0" PRIX32, syslog.stackFrame.pc);
-        LOG_FATAL("PSR = 0x%0" PRIX32, syslog.stackFrame.psr);
-
-        LOG_FATAL("* Fault Status Registers:");
-        LOG_FATAL("SCB->HFSR = 0x%0" PRIX32, syslog.registers.hfsr.all);
-        LOG_FATAL("SCB->CFSR = 0x%0" PRIX32, syslog.registers.cfsr.all);
-        LOG_FATAL("SCB->MMAR = 0x%0" PRIX32, syslog.registers.mmfar);
-        LOG_FATAL("SCB->BFAR = 0x%0" PRIX32, syslog.registers.bfar);
-        LOG_FATAL("SCB->ABFSR = 0x%0" PRIX32, syslog.registers.abfsr.all);
-        LOG_FATAL("EXC_RETURN = 0x%0" PRIX32, syslog.registers.exc_return);
-
-#if (DEBUG_DETAILED_HARD_FAULT_INFO == 1)
-        LOG_FATAL("* Details of Fault Status Registers:");
-        hfsr_t hfsr = syslog.registers.hfsr;
-        LOG_FATAL("SCB->HFSR = 0x%0" PRIX32 " - HardFault Status Register", hfsr.all);
-        if (hfsr.debugevt)
-            LOG_FATAL(" - DEBUGEVT: Hard fault caused by debug event");
-        if (hfsr.forced)
-            LOG_FATAL(" - FORCED: Hard fault caused by bus/memory management/usage fault");
-        if (hfsr.vecttbl)
-            LOG_FATAL(" - VECTBL: BusFault on vector table read");
-
-        cfsr_t cfsr = syslog.registers.cfsr;
-        LOG_FATAL("SCB->CFSR = 0x%0" PRIX32 " - Configurable Fault Status Register", cfsr.all);
-        mmfsr_t mmfsr = cfsr.mmfsr;
-        LOG_FATAL("> SCB->MMFSR = 0x%0" PRIX8 " Memory Management Fault Status Register", mmfsr.all);
-        if (mmfsr.mmarvalid)
-            LOG_FATAL(" - MMARVALID: MMAR is valid");
-        if (mmfsr.mlsperr)
-            LOG_FATAL(" - MLSPERR: A MemManage fault occurred during floating-point lazy state preservation");
-        if (mmfsr.mstkerr)
-            LOG_FATAL(" - MSTKERR: MemManage fault on stacking for exception entry");
-        if (mmfsr.munstkerr)
-            LOG_FATAL(" - MUNSTKERR: MemManage fault on unstacking for a return from exception:");
-        if (mmfsr.daccviol)
-            LOG_FATAL(" - DACCVIOL: Data access violation");
-        if (mmfsr.iaccviol)
-            LOG_FATAL(" - IACCVIOL: Instruction access violation");
-
-        bfsr_t bfsr = cfsr.bfsr;
-        LOG_FATAL("> SCB->BFSR = 0x%0" PRIX8 " - Bus Fault Status Register ", bfsr.all);
-        if (bfsr.bfarvalid)
-            LOG_FATAL(" - BFARVALID: BFAR is valid  ");
-        if (bfsr.lsperr)
-            LOG_FATAL(" - LSPERR: A bus fault occurred during floating-point lazy state preservation");
-        if (bfsr.stkerr)
-            LOG_FATAL(" - STKERR: Stacking error");
-        if (bfsr.unstkerr)
-            LOG_FATAL(" - UNSTKERR: Unstacking error");
-        if (bfsr.impreciserr)
-            LOG_FATAL(" - IMPREISERR: Imprecise data bus error");
-        if (bfsr.preciserr)
-            LOG_FATAL(" - PRECISERR: Precise data bus error");
-        if (bfsr.ibuserr)
-            LOG_FATAL(" - IBUSERR: Instruction bus error");
-
-        ufsr_t ufsr = cfsr.ufsr;
-        LOG_FATAL("> SCB->UFSR = 0x%0" PRIX16 " - Usage Fault Status Register", ufsr.all);
-        if (ufsr.divbyzero)
-            LOG_FATAL(" - DIVBYZERO: Divide by zero UsageFault");
-        if (ufsr.unaligned)
-            LOG_FATAL(" - UNALIGNED: Unaligned access UsageFault");
-        if (ufsr.nocp)
-            LOG_FATAL(" - NOCP: Attempt to execute a coprocessor instruction");
-        if (ufsr.invpc)
-            LOG_FATAL(" - INVPC: Invalid PC load UsageFault");
-        if (ufsr.invstate)
-            LOG_FATAL(" - INVSTATE: Invalid state UsageFault");
-        if (ufsr.undefinstr)
-            LOG_FATAL(" - UNDEFINSTR: Undefined instruction UsageFault");
-
-        abfsr_t abfsr = syslog.registers.abfsr;
-        LOG_FATAL("SCB->ABFSR = 0x%0" PRIX32 " - Auxiliary Bus Fault Status register", abfsr.all);
-        if (abfsr.itcm)
-            LOG_FATAL(" - ITCM: Asynchronous fault on ITCM interface");
-        if (abfsr.dtcm)
-            LOG_FATAL(" - DTCM: Asynchronous fault on DTCM interface.");
-        if (abfsr.ahbp)
-            LOG_FATAL(" - AHBP: Asynchronous fault on AHBP interface");
-        if (abfsr.eppb)
-            LOG_FATAL(" - EPPB: Asynchronous fault on EPPB interface");
-#endif // (DEBUG_DETAILED_HARD_FAULT_INFO == 1)
-#endif //(DEBUG_PRINT_HARD_FAULT_INFO == 1)
-    }
-
-    WEAK_AV void HardFault_Handler_C(syslog_exception_stack_frame_t *frame __attribute__((unused)),
-                                     uint32_t lr_value __attribute__((unused)))
-    {
-        static syslog_t syslog     = {};
-        syslog.stackFrame          = *frame;
-        syslog.registers.hfsr.all  = SCB->HFSR;
-        syslog.registers.cfsr.all  = SCB->CFSR;
-        syslog.registers.mmfar     = SCB->MMFAR;
-        syslog.registers.bfar      = SCB->BFAR;
-        syslog.registers.abfsr.all = SCB->ABFSR;
-
-        printHardFaultInfo(syslog);
-
-        abort();
-    }
-
-    WEAK_AV void MemManage_Handler_C(syslog_exception_stack_frame_t *frame __attribute__((unused)),
-                                     uint32_t lr __attribute__((unused)))
-    {
-        abort();
-    }
-
-#if defined(__cplusplus)
-}
-#endif
-
-WEAK_AV void MemManage_Handler(void)
-{
-#if 1
-    asm volatile(" tst lr,#4       \n"
-                 " ite eq          \n"
-                 " mrseq r0,msp    \n"
-                 " mrsne r0,psp    \n"
-                 " mov r1,lr       \n"
-                 " ldr r2,=MemManage_Handler_C \n"
-                 " bx r2"
-
-                 : /* Outputs */
-                 : /* Inputs */
-                 : /* Clobbers */
-    );
-#endif
-}
-
-[[gnu::naked]] WEAK_AV void HardFault_Handler(void)
-{
-    asm volatile(" tst lr,#4       \n"
-                 " ite eq          \n"
-                 " mrseq r0,msp    \n"
-                 " mrsne r0,psp    \n"
-                 " mov r1,lr       \n"
-                 " ldr r2,=HardFault_Handler_C \n"
-                 " bx r2"
-
-                 : /* Outputs */
-                 : /* Inputs */
-                 : /* Clobbers */
-    );
-}
-
-WEAK_AV void BusFault_Handler(void)
-{
-
-#if 0
-  asm volatile(
-      " tst lr,#4       \n"
-      " ite eq          \n"
-      " mrseq r0,msp    \n"
-      " mrsne r0,psp    \n"
-      " mov r1,lr       \n"
-      " ldr r2,=BusFault_Handler_C \n"
-      " bx r2"
-
-      : /* Outputs */
-      : /* Inputs */
-      : /* Clobbers */
-  );
-#endif
-}
-
-WEAK_AV void UsageFault_Handler(void)
-{
-
-#if 0
-  asm volatile(
-      " tst lr,#4       \n"
-      " ite eq          \n"
-      " mrseq r0,msp    \n"
-      " mrsne r0,psp    \n"
-      " mov r1,lr       \n"
-      " ldr r2,=UsageFault_Handler_C \n"
-      " bx r2"
-
-      : /* Outputs */
-      : /* Inputs */
-      : /* Clobbers */
-  );
-#endif
 }
 
 //*****************************************************************************
